+++ lib/json_schema_utils.dart
@@ -6,3 +6,4 @@ export 'src/json_schema.dart' show JsonSchema;
 export 'src/json_schema_document.dart' show JsonSchemaDocument;
 export 'src/json_type.dart' show JsonType;
 export 'src/string_json_schema.dart' show StringJsonSchema;
+export 'src/boolean_json_schema.dart' show BooleanJsonSchema;